# Conflicts:
#	build.gradle
#	gradle.properties
#	gradle/wrapper/gradle-wrapper.properties
#	src/main/java/omtteam/openmodularturrets/tileentity/turrets/AbstractDirectedTurret.java
#	src/main/java/omtteam/openmodularturrets/tileentity/turrets/LaserTurretTileEntity.java
#	src/main/java/omtteam/openmodularturrets/tileentity/turrets/ProjectileTurret.java
#	src/main/java/omtteam/openmodularturrets/tileentity/turrets/RayTracingTurret.java
#	src/main/java/omtteam/openmodularturrets/tileentity/turrets/TurretHead.java
#	src/main/java/omtteam/openmodularturrets/turret/TurretHeadUtil.java
#	src/test/java/omtteam/openmodularturrets/util/OMTUtilTest.java

+++ src/main/java/omtteam/openmodularturrets/tileentity/turrets/RailGunTurretTileEntity.java
@@ -95,7 +95,9 @@ public class RailGunTurretTileEntity extends RayTracingTurret {
         if (!hitBlock.getMaterial().isSolid()) {
             return;
         } else if (OMTConfig.TURRETS.canRailgunDestroyBlocks) {
+            if(hitBlock.getBlockHardness(this.world,pos) < 200){
             this.getWorld().destroyBlock(pos, false);
+            }
         }
     }
 }
